requirements.txt
cryptography
rich
pillow
pytesseract
pyyaml
